pyproject.toml
[tool.poetry]
name = "parking_lot"
version = "0.1.0"
description = ""
authors = ["Jefte Keller <kellergnr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.3"
djangorestframework = "^3.12.4"
environs = "^9.3.2"
coverage = "^5.5"

[tool.poetry.dev-dependencies]
pylint = "^2.8.3"
yapf = "^0.31.0"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
